+++ app/client/packages/ui/components/__tests__/filter-bar.spec.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+import FilterBar from '../filter-bar.tsx';
+
+test('Filter bar', () => {
+    const component = mountWithTheme(<FilterBar filter="search" />);
+    expect(component).toMatchSnapshot();
+});
